@W-16589746: [Android] REST wrappers for select SFAP APIs (Consolidate `sfap_api` `embeddings` REST wrappers into a single client class)
